Modify spec tests to only test monitoring on systemd platforms, de-lint

+++ modules/daemon/spec/daemon-restart-on-kill/03-stop-sleep-start.pp
@@ -1,6 +1,15 @@
 node default {
 
-  exec { 'systemctl stop my-program && sleep 20 && systemctl start my-program':
-    path => ['/usr/local/sbin', '/usr/local/bin', '/usr/sbin', '/usr/bin', '/sbin', '/bin']
+  if ($::service_provider == 'debian') {
+    exec { 'noop':
+      command => '/bin/true',
+      path => ['/usr/local/sbin', '/usr/local/bin', '/usr/sbin', '/usr/bin', '/sbin', '/bin']
+    }
+  }
+
+  if ($::service_provider == 'systemd') {
+    exec { 'systemctl stop my-program && sleep 20 && systemctl start my-program':
+      path => ['/usr/local/sbin', '/usr/local/bin', '/usr/sbin', '/usr/bin', '/sbin', '/bin']
+    }
   }
 }
